+++ tests/Text.Format.tests.js
@@ -151,12 +151,12 @@ tools.test( function withAngleFontSize() {
 });
 
 
-tools.test( function inverse() {
-  expect(rac.Text.Format.topRight.inverse()).equalsTextFormat(ha.left, va.bottom, 1/2);
-  expect(rac.Text.Format.bottomLeft.inverse()).equalsTextFormat(ha.right, va.top, 1/2);
-  expect(rac.Text.Format.centerCenter.inverse()).equalsTextFormat(ha.center, va.center, 1/2);
+tools.test( function reverse() {
+  expect(rac.Text.Format.topRight.reverse()).equalsTextFormat(ha.left, va.bottom, 1/2);
+  expect(rac.Text.Format.bottomLeft.reverse()).equalsTextFormat(ha.right, va.top, 1/2);
+  expect(rac.Text.Format.centerCenter.reverse()).equalsTextFormat(ha.center, va.center, 1/2);
 
-  expect(upright.inverse()).equalsTextFormat(ha.right, va.baseline, 1/4);
+  expect(upright.reverse()).equalsTextFormat(ha.right, va.baseline, 1/4);
 });