+++ changelog.md
@@ -18,6 +18,7 @@ Changelog
 + Modify `point.text` to allow an optional `format` parameter
 + Rename `Text.Format.horizontal` to `horizontalAlign`
 + Rename `Text.Format.vertical` to `verticalAlign`
++ Rename `text.format.inverse` to `reverse`
 + Document `P5Drawer.setDrawFunction`, `P5Drawer.setDebugFunction`, `Rac.utils`, `rac.utils`